refactor(default-icons): Rename file to platform icon sizes
