+++ src/platform-icon-sizes.js
